requirements.txt
openai
requests
beautifulsoup4
lxml
python-dotenv
